# Conflicts:
#	src/main/java/com/example/rewrite/controller/ProdController.java
#	src/main/java/com/example/rewrite/service/user/UserService.java

+++ src/main/java/com/example/rewrite/service/prod/ProdService.java
@@ -12,7 +12,7 @@ public interface ProdService {
     List<ProductDTO> getAllProducts();
 
     // 상품 상세 조회
-    ProductDTO getProductById(int id);
+    ProductDTO getProductById(Long id);
 
     // 상품 등록
     ProductDTO registerProduct(ProductDTO productDTO);
